fix: P9-P13 audit — 16 fixes across 12 source files
Critical:
- CC-C1: Chunked Hessian in fno_greeks.py (5GB→1.5GB VRAM, prevents OOM)
- P10-C2: SDE simulation float32→float64 in portfolio_mc.py
- P13-C1: Signature SDE depth/dim combinatorial explosion guard
- P13-C2: RKHS call-put parity with explicit discount factor
- P9-I1: RKHS variance clamp 1e-8→1e-4 (Durrleman compliance)

Warning:
- CC-W1: ConflatedQueue lost-wakeup race (event.clear inside lock)
- CC-I1: orjson fallback warning log in websocket.py
- P11-W1: BSDE gradient clipping (max_norm=10) in indifference_pricing.py
- P13-W1: EGNO correlation PSD validation via spectral clamping
- P13-W3: RKHS per-call Generator instead of global manual_seed(42)
- P10-W1: Centralized strike resolver in utils/strikes.py + 4 call sites

Info:
- P11-I1: Explicit precompute attribute in barrier_hedging.py
- P12-I1: Parameterized workspace_gb in compile_trt.py
- P13-I1: Parameterized N_paths in rkhs_mlsv.py price_option()

Test results: 953 passed, 2 skipped, 265 warnings in 497.96s

+++ src/deepvol/risk/portfolio_mc.py
@@ -110,16 +110,16 @@ class MonteCarloVaREngine:
         num_blocks = (N_paths + block_size - 1) // block_size
         PATHS_padded = num_blocks * block_size
         
-        # Structure of Arrays (SoA) layout
-        S_t = torch.full((PATHS_padded,), S0, dtype=torch.float32, device=self.device)
-        V_t = torch.full((PATHS_padded,), v0, dtype=torch.float32, device=self.device)
+        # Structure of Arrays (SoA) layout — float64 per AGENTS.md for SDE pricing layers
+        S_t = torch.full((PATHS_padded,), S0, dtype=torch.float64, device=self.device)
+        V_t = torch.full((PATHS_padded,), v0, dtype=torch.float64, device=self.device)
         log_S = torch.log(S_t)
 
         delta_t = dt / N_steps
         total_rows = num_blocks * 2 * N_steps
 
         # Generate normal random numbers and lay them out in block-tiled shape [N, B]
-        raw_randn = torch.randn((2 * N_steps, PATHS_padded), dtype=torch.float32, device=self.device)
+        raw_randn = torch.randn((2 * N_steps, PATHS_padded), dtype=torch.float64, device=self.device)
         reshaped = raw_randn.view(2 * N_steps, num_blocks, block_size)
         permuted = reshaped.permute(1, 0, 2)  # [num_blocks, 2 * N_steps, block_size]
         block_tiled = permuted.reshape(total_rows, block_size).contiguous()
